pyproject.toml
[tool.poetry]
name = "mask-detection"
version = "0.1.0"
description = ""
authors = ["viniciusrsouza <vinicius.rsouza07@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.*"
argparse = "1.1"
imutils = "0.5.3"
keras = "2.3.1"
matplotlib = "3.2.1"
numpy = "1.19.2"
opencv-python = "4.2.0.*"
pillow="8.2"
scipy = "1.4.1"
tensorflow = "^2.4.2"
tqdm = "^4.61.1"
xmltodict = "0.12.0"
keras-tuner = "^1.0.3"
scikit-learn = "^0.24.2"

[tool.poetry.dev-dependencies]
autopep8 = "1.5.7"
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
